// lib/api/generated/model/planDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */
import type { PlanFeatureDto } from "./planFeatureDto";

export interface PlanDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  monthlyPrice?: number;
  yearlyPrice?: number;
  lifetimePrice?: number;
  isActive?: boolean;
  /** @nullable */
  planFeatures?: PlanFeatureDto[] | null;
}